Move draw call to Instance
Also add some notes about error handling for adding VBO data
